Replaced all keyword representing 'int' values by 'byte' because there are less than 256 different token kinds .. a parser optimization?

+++ DParser2/Completion/ICompletionDataGenerator.cs
@@ -7,7 +7,7 @@ namespace D_Parser.Completion
 		/// <summary>
 		/// Adds a token entry
 		/// </summary>
-		void Add(int Token);
+		void Add(byte Token);
 
 		/// <summary>
 		/// Adds a property attribute
